sharedobjcat: allow addition/deletion of object in same batch
On rare occasions, if we cancel a compaction at a point right after
creating a new object, we'll end up deleting the object shortly after.
Sometimes the two can happen with no explicit Sync() call in between,
resulting in the object being created and deleted in the same batch
which currently results in an error. This change allows for this case
by moving around logic in the catalog batch application code.